Refactor key management algorithms into own namespace
