+++ src/SimpleJWT/Crypt/KeyManagement/KeyEncryptionAlgorithm.php
@@ -33,8 +33,9 @@
  * IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace SimpleJWT\Crypt;
+namespace SimpleJWT\Crypt\KeyManagement;
 
+use SimpleJWT\Crypt\CryptException;
 use SimpleJWT\Keys\KeySet;
 use SimpleJWT\Keys\KeyException;
